Rename files by interface
