+++ src/go/cmd/root/cli_config_parser.go
@@ -2,9 +2,7 @@ package cmdroot
 
 import (
 	"bufio"
-	"fmt"
 	"io"
-	"net/url"
 	"strings"
 
 	"github.com/kkrull/marmot/core"
@@ -106,86 +104,4 @@ func (parser rootConfigParser) ParseR(
 
 		return config, nil
 	}
-}
-
-// Application configuration derived from flags passed to the CLI.
-type rootCliConfig struct {
-	// Application interface
-	cmdFactory   use.CommandFactory
-	queryFactory use.QueryFactory
-
-	// CLI arguments
-	args []string
-
-	// CLI flags
-	debug        bool
-	flagSet      *pflag.FlagSet
-	metaRepoPath string
-
-	// CLI input
-	inputLines []string
-}
-
-/* Application interface */
-
-func (params rootCliConfig) CommandFactory() use.CommandFactory { return params.cmdFactory }
-func (params rootCliConfig) QueryFactory() use.QueryFactory     { return params.queryFactory }
-
-/* CLI arguments */
-
-func (params rootCliConfig) Args() []string { return params.args }
-
-func (params rootCliConfig) ArgsAsUrls() ([]*url.URL, error) {
-	urls := make([]*url.URL, len(params.args))
-	for i, rawArg := range params.args {
-		if urlArg, parseErr := url.Parse(rawArg); parseErr != nil {
-			return nil, fmt.Errorf("url expected: <%s>; %w", rawArg, parseErr)
-		} else {
-			urls[i] = urlArg
-		}
-	}
-
-	return urls, nil
-}
-
-/* CLI debugging */
-
-func (params rootCliConfig) Debug() bool { return params.debug }
-
-func (params rootCliConfig) PrintDebug(writer io.Writer) {
-	for i, arg := range params.args {
-		fmt.Fprintf(writer, "arg [%d]: %s\n", i, arg)
-	}
-
-	for _, flag := range rootFlags {
-		fmt.Fprintf(writer, "flag --%s=%s\n", flag.LongName(), flag.Find(params.flagSet))
-	}
-
-	for i, line := range params.inputLines {
-		fmt.Fprintf(writer, "stdin [%d]: %s\n", i, line)
-	}
-}
-
-/* CLI flags */
-
-func (params rootCliConfig) MetaRepoPath() string { return params.metaRepoPath }
-
-/* CLI input */
-
-func (params rootCliConfig) InputLines() []string { return params.inputLines }
-
-func (params rootCliConfig) InputLinesAsUrls() ([]*url.URL, error) {
-	urls := make([]*url.URL, 0)
-	for _, rawLine := range params.inputLines {
-		trimmedLine := strings.TrimSpace(rawLine)
-		if trimmedLine == "" {
-			continue
-		} else if urlLine, parseErr := url.Parse(rawLine); parseErr != nil {
-			return nil, fmt.Errorf("url expected: <%s>; %w", rawLine, parseErr)
-		} else {
-			urls = append(urls, urlLine)
-		}
-	}
-
-	return urls, nil
 }
