+++ src/go/cmd/root/root_flagset.go
@@ -2,15 +2,15 @@ package cmdroot
 
 import "github.com/spf13/cobra"
 
-// Flag configuration for the root (e.g. top-level) command that dispatches to all other commands.
-func RootFlagSet() CommandFlagSet {
-	return &rootFlagSet{}
-}
-
 // Flags that can be passed to a CLI command.
 type CommandFlagSet interface {
 	// Add the implemented flags to the given CLI command.
 	AddTo(cmd *cobra.Command) error
+}
+
+// Flag configuration for the root (e.g. top-level) command that dispatches to all other commands.
+func RootFlagSet() CommandFlagSet {
+	return &rootFlagSet{}
 }
 
 type rootFlagSet struct{}
